refactor(swap_encryption): address all Zac review comments
