+++ perfkitbenchmarker/data/cluster/swap_encryption_daemonset.yaml.j2
@@ -1,3 +1,62 @@
+---
+# ConfigMap holds the setup script so it can be reviewed and modified
+# independently of the DaemonSet manifest structure.
+apiVersion: v1
+kind: ConfigMap
+metadata:
+  name: {{ ds_name }}-setup
+  namespace: {{ ds_namespace }}
+data:
+  setup.sh: |
+    #!/usr/bin/env bash
+    # swap_encryption_setup.sh — privileged DaemonSet init for PKB swap benchmarks.
+    # Installs measurement tools, verifies the swap device, then writes /tmp/pkb_ready.
+    set -euo pipefail
+
+    echo "[pkb] Installing measurement tools..."
+    # Only tools needed for Phase 1 (raw-device fio) and Phase 2 (CPU/I/O overhead).
+    # Workload benchmarks (redis, opensearch, kernel-build) run in separate pods.
+    PKB_APT_OK=0
+    for _attempt in 1 2 3; do
+      DEBIAN_FRONTEND=noninteractive apt-get update -qq 2>&1 || true
+      DEBIAN_FRONTEND=noninteractive apt-get install -y -qq \
+        fio \
+        cryptsetup \
+        mdadm \
+        sysstat \
+        nvme-cli \
+        2>&1 && PKB_APT_OK=1 && break
+      echo "[pkb] apt-get attempt $_attempt failed, retrying in 15s..." >&2
+      sleep 15
+    done
+    if [ "$PKB_APT_OK" != "1" ] || ! command -v fio >/dev/null 2>&1; then
+      echo "[pkb] FATAL: fio not installed after 3 attempts" >&2
+      exit 1
+    fi
+    echo "[pkb] fio installed: $(fio --version 2>&1 | head -1)"
+
+    echo "[pkb] Verifying swap device is active..."
+    PKB_SWAP_FOUND=0
+    for _attempt in $(seq 1 30); do
+      if awk 'NR>1{found=1} END{exit !found}' /proc/swaps 2>/dev/null; then
+        PKB_SWAP_DEV=$(awk 'NR==2{print $1}' /proc/swaps)
+        echo "[pkb] Swap device active: $PKB_SWAP_DEV"
+        PKB_SWAP_FOUND=1
+        break
+      fi
+      echo "[pkb] Waiting for swap device (attempt $_attempt/30)..." >&2
+      sleep 5
+    done
+    if [ "$PKB_SWAP_FOUND" != "1" ]; then
+      echo "[pkb] FATAL: no active swap device after 150s." \
+           "Check linuxConfig.swapConfig / kubelet swap config." >&2
+      exit 1
+    fi
+
+    echo "[pkb] Measurement tools ready. Writing ready sentinel."
+    touch /tmp/pkb_ready
+    exec sleep infinity
+---
 apiVersion: apps/v1
 kind: DaemonSet
 metadata:
@@ -26,50 +85,7 @@ spec:
         image: {{ image }}
         command:
         - bash
-        - -c
-        - |
-          echo "[pkb] Installing measurement tools..."
-          # Only the tools needed for Phase 1 (raw-device fio) and Phase 2
-          # (CPU/I/O overhead) are installed here.  Workload benchmarks
-          # (redis, opensearch, kernel-build) run in separate pods via
-          # existing PKB benchmark modules and are NOT installed here.
-          PKB_APT_OK=0
-          for _attempt in 1 2 3; do
-            apt-get update -qq 2>&1 || true
-            DEBIAN_FRONTEND=noninteractive apt-get install -y -qq \
-              fio \
-              cryptsetup \
-              mdadm \
-              sysstat \
-              nvme-cli \
-              2>&1 && PKB_APT_OK=1 && break
-            echo "[pkb] apt-get attempt $_attempt failed, retrying in 15s..." >&2
-            sleep 15
-          done
-          if [ "$PKB_APT_OK" != "1" ] || ! command -v fio >/dev/null 2>&1; then
-            echo "[pkb] FATAL: fio not installed after 3 attempts" >&2
-            exit 1
-          fi
-          echo "[pkb] fio installed: $(fio --version 2>&1 | head -1)"
-          echo "[pkb] Verifying swap device is active..."
-          PKB_SWAP_FOUND=0
-          for _attempt in $(seq 1 30); do
-            if awk 'NR>1{found=1} END{exit !found}' /proc/swaps 2>/dev/null; then
-              PKB_SWAP_DEV=$(awk 'NR==2{print $1}' /proc/swaps)
-              echo "[pkb] Swap device active: $PKB_SWAP_DEV"
-              PKB_SWAP_FOUND=1
-              break
-            fi
-            echo "[pkb] Waiting for swap device (attempt $_attempt/30)..." >&2
-            sleep 5
-          done
-          if [ "$PKB_SWAP_FOUND" != "1" ]; then
-            echo "[pkb] WARNING: no active swap device after 150s — " \
-                 "check linuxConfig.swapConfig / kubelet swap config." >&2
-          fi
-          echo "[pkb] Measurement tools ready. Writing ready sentinel."
-          touch /tmp/pkb_ready
-          sleep infinity
+        - /scripts/setup.sh
         securityContext:
           privileged: true
           capabilities:
@@ -83,6 +99,9 @@ spec:
             fieldRef:
               fieldPath: spec.nodeName
         volumeMounts:
+        - name: scripts
+          mountPath: /scripts
+          readOnly: true
         - name: dev
           mountPath: /dev
         - name: sys
@@ -98,6 +117,10 @@ spec:
           mountPath: /lib/modules
           readOnly: true
       volumes:
+      - name: scripts
+        configMap:
+          name: {{ ds_name }}-setup
+          defaultMode: 0755
       - name: dev
         hostPath:
           path: /dev
