+++ perfkitbenchmarker/providers/gcp/google_kubernetes_engine.py
@@ -37,7 +37,7 @@ from perfkitbenchmarker.resources.container_service import container_registry
 from perfkitbenchmarker.resources.container_service import kubectl
 from perfkitbenchmarker.resources.container_service import kubernetes_cluster
 from perfkitbenchmarker.resources.container_service import kubernetes_commands
-from perfkitbenchmarker.resources.container_service import swap_config as swap_config_lib
+from perfkitbenchmarker.providers.gcp import gcp_swap_config as gcp_swap_config_lib
 
 FLAGS = flags.FLAGS
 
@@ -447,14 +447,35 @@ class GkeCluster(BaseGkeCluster):
       )
       self._AddNodeParamsToCmd(nodepool, cmd)
       # If swap_config wrote a linuxConfig tempfile, clean it up after Issue().
-      swap_cfg = getattr(nodepool, '_gke_swap_config', None)
+      swap_cfg = getattr(nodepool, 'gke_swap_config', None)
       try:
         self._IssueResourceCreationCommand(cmd)
       finally:
         if swap_cfg is not None:
           swap_cfg.CleanupYaml()
-          nodepool._gke_swap_config = None  # pylint: disable=protected-access
+          nodepool.gke_swap_config = None
       self._CreateCustomComputeClass(nodepool)
+
+  def DeleteDefaultNodePool(self, name: str = 'default-pool') -> None:
+    """Delete a nodepool by name (typically the dummy e2-medium default-pool).
+
+    GKE requires at least one nodepool at cluster creation time. The default
+    pool can be deleted once the benchmark nodepool and its DaemonSet are
+    Running. Must be called from the benchmark, not _GcloudCommand directly.
+
+    Args:
+      name: Nodepool name to delete (default: 'default-pool').
+
+    Raises:
+      errors.VmUtil.IssueCommandError: if the gcloud command returns non-zero.
+    """
+    cmd = self._GcloudCommand(
+        'container', 'node-pools', 'delete', name,
+        '--cluster', self.name,
+    )
+    cmd.args.append('--quiet')
+    logging.info('[gke] Deleting nodepool: %s', name)
+    cmd.Issue(timeout=300)
 
   def _CreateCustomComputeClass(
       self, nodepool_config: container.BaseNodePoolConfig
@@ -604,10 +625,12 @@ class GkeCluster(BaseGkeCluster):
 
     # Per-nodepool swap config takes precedence over the global flag.
     if nodepool_config.swap_config is not None:
-      gke_swap = swap_config_lib.GkeSwapConfig.from_spec(nodepool_config.swap_config)
+      gke_swap = gcp_swap_config_lib.GkeSwapConfig.from_spec(
+          nodepool_config.swap_config
+      )
       cmd.flags['system-config-from-file'] = gke_swap.WriteLinuxConfigYaml()
       # Store on nodepool so _CreateNodePools() can clean up the tempfile.
-      nodepool_config._gke_swap_config = gke_swap  # pylint: disable=protected-access
+      nodepool_config.gke_swap_config = gke_swap
       # dm-crypt requires UBUNTU_CONTAINERD (Ajay r3472549985).
       cmd.flags['image-type'] = 'UBUNTU_CONTAINERD'
       # Prevent GKE from replacing the node after swap setup is complete.
