+++ perfkitbenchmarker/resources/container_service/container.py
@@ -191,6 +191,10 @@ class BaseNodePoolConfig:
     # declares swap_config. Consumed by _AddNodeParamsToCmd() in the cloud
     # provider to apply swap configuration during nodepool creation.
     self.swap_config: container_spec_lib.SwapConfigSpec | None = None
+    # Written by GkeCluster._AddNodeParamsToCmd() after building the GKE
+    # linuxConfig YAML tempfile. Cleared by _CreateNodePools() after Issue().
+    # Type: gcp_swap_config.GkeSwapConfig | None (avoids circular import here).
+    self.gke_swap_config = None
     self.max_local_disks: int | None
     self.ssd_interface: str | None
     self.threads_per_core: int
